third_party/WebKit/LayoutTests/http/tests/security/powerfulFeatureRestrictions/webshare-on-insecure-origin.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script>
if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
  window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
} else {
  test(() => {
    assert_false('share' in navigator, 'navigator has attribute \'share\'.');
  }, 'navigator.share should be undefined in non-secure context');
}
</script>
